requirements.txt
pymupdf>=1.23.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
